+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Merged SFR-1415 branch with old branch
 - Tugboat deployment for SFR-1415 branch commmit
 - Formatting for pubDateCheck tests methods
+- Importing datetime objects in test_sfrRecord manager
 
 ## 2021-11-22 -- v0.9.5
 ### Added